A More Reliable Keyframe Command
The new command asks ffprobe for a list of all frames, but with an extra flag on each one telling us if it's a keyframe (key_frame=1) or not (key_frame=0). Our Python code will then simply filter that list to get the timestamps for only the keyframes. This method is faster and much more compatible with a wider range of video files.